backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.14
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
python-dotenv==1.0.0
slowapi==0.1.9
refuel-autolabel[openai]>=0.0.16
openai==0.28.1
pandas>=2.1.3 